+++ tests/integrations/torch/optim_test.py
@@ -2,8 +2,8 @@ from tango.integrations.torch.optim import Optimizer, LRScheduler
 
 
 def test_all_optimizers_registered():
-    assert "Adagrad" in Optimizer.list_available()
+    assert "torch::Adagrad" in Optimizer.list_available()
 
 
 def test_all_lr_schedulers_registered():
-    assert "ExponentialLR" in LRScheduler.list_available()
+    assert "torch::ExponentialLR" in LRScheduler.list_available()